Move typelizer fix for flaky tests to test env
